insurance-adk/requirements.txt
# Official Google ADK (now available!)
google-adk>=1.0.0

# LLM Integration  
litellm>=1.55.0
openai>=1.80.0
anthropic>=0.50.0

# FastAPI Server Integration
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.11.0

# Core Dependencies
httpx>=0.28.0
aiohttp>=3.12.0
pyyaml>=6.0.2
python-dotenv>=1.1.0

# Monitoring (Preserved from current system)
langfuse>=2.60.0
prometheus-client>=0.22.0
structlog>=25.3.0

# MCP Integration (Preserved from current system)
fastmcp>=2.5.0
python-socketio>=5.13.0

# Testing
pytest>=8.3.0
pytest-asyncio>=1.0.0
pytest-cov>=6.1.0
pytest-mock>=3.14.0

# Development
black>=25.1.0
flake8>=7.2.0
mypy>=1.16.0 